web, core: fix grammatical issue in stage bindings
Global search & replace the phrase "...when the Stage is present
to the user." with "...when the Stage is presented to the user."

Because it's bad grammar.  Since the source language phrase has
been changed in all cases, it will have the same hash and the
resulting translations will still be targeted the same.