web: exercise: deconstruct Table
Create four variants of @beryju's table, none of which talk to the API:

- ak-simple-table: Just a table.  Sortable
- ak-select-table: A table with a "select" column. Sortable. Handles `multiple` (multiselect
  protocol), is single (radio protocol) by default. Emits `change` on, well, change.
- ak-expandable-table: A table with an 'expansion' section per row.
- ak-table-view: A table with both `select` and `expansion` sections.

All the various components are overridable and inheritable, as in the original, but these tables are
quite usable in their current form without doing so.

As an extension of @beryju's table, they also have (ridiculously comprehensive) `part::` syntax, so
every element that you might want to mangle with CSS is accessible.

The tables take, at a minimum, `string[][]`, but have much richer types, leaning on the "Flat" and
"Group" collections similar to those used by SearchSelect, DualSelect, and ListSelect, but with more
complex content fields. Expansion tables take, per row, a field named and typed `expansion?: () =>
TemplateResult`; clients are expected to fill in the blanks.

Selectable tables have public accessors `value` (a string representation of the keys, separated by a
field separator that can be specified), `selected` (returns the array of keys), and `json()`, which
also returns the array of keys; the last conforms to the AKDataComponent protocol.

Selectable tables emit `change` onChange; as is standard, it is up to the client to access the
`event.target` and extract the values it wants.

This started as an attempt to permit us to have static tables on the client side. The `Table` class
supplied by authentik is deeply integrated with the API. It kinda spiraled out of control. I haven't
decided if that's glorious or unfortunate yet.

- Stories for the expansion tables
- Unit tests for all the tables.
- Independent pagination implementation (see ak-dual-select-pagination)
- Independent toolbar implementation (see ak-dual-select-controls)
- Independent chip implementation
- API Integration
- Experiment with replacing existing tables with this beast.

Replace the TablePage with a more compositional variant, allowing
us to standardize the View class, which currently has a lot of
variants.