web: fix navigation issues inside wizard
- Replace index-based steps with name-based step management
- Rename a _lot_ of components to have shorter, more straightforward names
- Rename files to match those of their components
- Remove default exports; prefer fully named exports
- Create child events rather than custom events; prefer type-checked event handling
- Move step declarations into the same folder with the step renderer
- Steps are now defined in the wizard directly
- Re-arrange some classes to meet with our class definition pattern
- Rename event handler to use the `on...` prefix
- Replace some switch/case with `ts-pattern`
- Migrate `WizardController` functionality into Wizard base class

The first change is the most essential: a wizard is a _state machine_, in which each page is both
dialogue with the user and a validation pass on what page to pass control to next (or permit to go
back). By using name-based step management we enable the state machine to branch, which in turn
permits loops allowing collections to be created inside the wizard. This is an essential step to
enabling User/Group/Policy bindings (note the plural) in the Application Wizard.

Using subclasses of `Event` instead of `CustomEvent` permits stronger type checking. It's considered the
"correct" way.

Forbidding default exports enforces consistency of names. A rename now requires a deliberate and
explicit step.

For single-component files, the filename should always match the component.  For single classes, the
same rules applies.

Some of the names were naive, overly long, or could misinform. I've made an effort to fix those
based on my knowledge of the authentik service as it now stands.

Moving each step's managerial class into the same folder with the renderer means that the "what to
draw" and "how to draw" are in the same place.  No more digging in two seperate folders for
information on how a page is managed and rendered.

Using `ts-pattern` enables stricter type checking at compile time.

With the string-based step system, `WizardController` was so small it no longer seemed viable to
have the controls and the state in two separate classes. It was originally thought that
`WizardController` would have uses outside of the wizard itself, but that hasn't panned out.

The rest is lint and code quality stuff.