web: fix navigation issues inside wizard
- Replace index-based steps with name-based step management
- Steps are now declared in the wizard directly
- Migrate `WizardController` functionality into Wizard base class

The first change is the most essential: a wizard is a _state machine_, in which each page is both
dialogue with the user and a validation pass on what page to pass control to next (or permit to go
back). By using name-based step management we enable the state machine to branch, which in turn
permits loops allowing collections to be created inside the wizard. This is an essential step to
enabling User/Group/Policy bindings (note the plural) in the Application Wizard.

Using subclasses of `Event` instead of `CustomEvent` permits stronger type checking. It's considered the
"correct" way.

With the string-based step system, `WizardController` was so small it no longer seemed viable to
have the controls and the state in two separate classes. It was originally thought that
`WizardController` would have uses outside of the wizard itself, but that hasn't panned out.

The rest is lint and code quality stuff.