web: Add Policy/User/Group Bindings to the Application Wizard
The navigation is in place, but is not yet completely tested. The interaction is a little broken,
because you can possibly save an incomplete binding. I haven't finished revising the validator, but
I'm out of time on a Friday night, and I need to get to class.  The real trick will be intercepting
the change events and recording them independently.  I'd like to do that without having to
reengineer the wizard any more, thank you.

Also: Fixed a bug in search-select-ez that would set the groupBy incorrectly if you didn't provide
one yourself.